+++ client.js
@@ -18,6 +18,17 @@ export const PiholeClient = GObject.registerClass(
       this._versionsUrl = this._authUrl + "&versions";
 
       this._session = new Soup.Session();
+      this._session.set_user_agent(`Phi/2.0 `);
+      this._decoder = new TextDecoder();
+
+      this.data = {};
+      this.data.dns_queries_today = "Initializing";
+      this.data.ads_blocked_today = "Initializing";
+      this.data.ads_percentage_today = "Initializing";
+      this.data.domains_being_blocked = "Initializing";
+      this.data.version = "Initializing";
+      this.data.blocking = false;
+      this.data.updateExists = false;
     }
 
     async _readAsString(input_stream) {
@@ -32,12 +43,12 @@ export const PiholeClient = GObject.registerClass(
       );
 
       const bytes = output_stream.steal_as_bytes();
-      const text_decoder = new TextDecoder("utf-8");
-      return text_decoder.decode(bytes.toArray().buffer);
+      return this._decoder.decode(bytes.toArray());
     }
 
     async _fetchUrl(url) {
       const message = Soup.Message.new("GET", url);
+
       const input_stream = await this._session.send_async(
         message,
         GLib.PRIORITY_DEFAULT,
@@ -46,7 +57,7 @@ export const PiholeClient = GObject.registerClass(
 
       if (message.status_code !== Soup.Status.OK) {
         this._handleError(ClientError.HTTP_NOT_OK);
-        return;
+        return JSON.parse("{}");
       }
 
       const data = await this._readAsString(input_stream);
@@ -55,12 +66,26 @@ export const PiholeClient = GObject.registerClass(
 
     async fetchSummary() {
       const json = await this._fetchUrl(this._summaryUrl);
-      return json;
+      this.data.dns_queries_today = json.dns_queries_today;
+      this.data.ads_blocked_today = json.ads_blocked_today;
+      this.data.ads_percentage_today = json.ads_percentage_today;
+      this.data.domains_being_blocked = json.domains_being_blocked;
+
+      // Status: enabled/disabled
+      this.data.blocking = json.status;
     }
 
     async fetchVersion() {
       const json = await this._fetchUrl(this._versionsUrl);
-      return json;
+      this.data.version = json.core_current;
+
+      this.data.updateExists =
+        json.core_update || json.web_update || json.FTL_update;
+    }
+
+    async fetchData() {
+      await this.fetchSummary();
+      await this.fetchVersion();
     }
 
     togglePihole(state) {
@@ -69,6 +94,7 @@ export const PiholeClient = GObject.registerClass(
         : this._authUrl + "&disable";
 
       this._fetchUrl(toggleUrl).catch();
+      this.data.blocking = state ? "enabled" : "disabled";
     }
 
     destroy() {