+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v2.0
+
+- Support Pi-hole v6. Fixes [#10].
+- Fetch sensor data: CPU, memory usage and temperature. This feature requires Pi-hole v6. Fixes [#9].
+- Use appropriate widgets for password/token entry.
+
 ## v1.8
 
 - Add option to disable checking Pi-hole software updates.